Added website files
Yes, I know, I should have worked on it on top of git, but I had not Internet connection while I was working, so it is what it is.

This is a commit with all of the files from both the template and the kinda finished site.

TODO : Fix discord link and some final touches I suppose.