Added placeholder image for image loading
When an image has not finished loading, the src of an <img> tag cannot be changed on the screen. One can set it on the code, but it will not change on the screen until the requested target image has finished loading. Rather than setting the selected default image for the image popup to a large screenshot, it can be set to be this placeholder which loads fast.